Fix specs and Cache-Control bug for Rails 4
