+++ spec/cache_control_manager_spec.rb
@@ -1,42 +1,15 @@
-require 'heroku-deflater/cache_control_manager'
-require 'rails'
+require 'spec_helper'
 
 describe HerokuDeflater::CacheControlManager do
-  class Rails5App
-    def config
-      @_config ||= Config.new
-    end
-
-    class Config
-      attr_accessor :headers
-
-      def initialize
-        @headers = {}
-      end
-
-      def public_file_server
-        self
-      end
-    end
-  end
-
-  class Rails4App
-    def config
-      @_config ||= Config.new
-    end
-
-    class Config
-      attr_accessor :static_cache_control
-    end
-  end
+  subject { described_class.new(app) }
+  before { subject.setup_max_age(86400) }
 
   context 'Rails 4.x and below' do
     let(:app) { Rails4App.new }
 
-    subject { described_class.new(app) }
     before do
-      allow(subject).to receive(:rails_version_5?).and_return(false)
-      subject.setup_max_age(86400)
+      allow(HerokuDeflater).to receive(:rails_version_5?).and_return(false)
+      subject.cache_control_headers
     end
 
     it 'sets max age for static_cache_control config option' do
@@ -51,10 +24,9 @@ describe HerokuDeflater::CacheControlManager do
   context 'Rails 5' do
     let(:app) { Rails5App.new }
 
-    subject { described_class.new(app) }
     before do
-      allow(subject).to receive(:rails_version_5?).and_return(true)
-      subject.setup_max_age(86400)
+      allow(HerokuDeflater).to receive(:rails_version_5?).and_return(true)
+      subject.cache_control_headers
     end
 
     it 'sets max age for public_file_server config option' do
